🚚Rename gatsby-api to render-helpers
