+++ gatsby-theme-fela/gatsby-browser.js
@@ -1,5 +1,5 @@
 const React = require("react")
-const { wrapWithFelaRenderer } = require("./gatsby-api")
+const { wrapWithFelaRenderer } = require("./render-helpers")
 const FelaProvider = require("./src/fela/FelaProvider").default
 
 exports.wrapRootElement = ({ element }) => {
